Fix build in pciutils
Three CLs are merged in the same topic as ag/28992874, but
ag/28992964 was skipped as it already exists on main.

The code inconsistency comes from this merge conflict resolution
(ag/28992974), which should have been fixed by merging relevant
CLs in the correct order instead.

This CL is a reland of ag/28217062 with different change ID.

Bug: 362369388
Bug: http://b/351833436
Test: m
Change-Id: I1c34db1f6ce330b7661675eaf2dbbb4c1f169a18